# benchmark: performance of running ollama server